refactor(MenuDrawerItemView): Move Items view to new package
