+++ app/src/main/java/com/infomaniak/mail/views/decoratedTextItemView/DecoratedTextItemView.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.views
+package com.infomaniak.mail.views.decoratedTextItemView
 
 import android.content.Context
 import android.graphics.drawable.Drawable
