+++ app/src/main/java/com/infomaniak/mail/views/decoratedTextItemView/SimpleFolderItemView.kt
@@ -15,13 +15,24 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.views
+package com.infomaniak.mail.views.decoratedTextItemView
 
 import android.content.Context
 import android.util.AttributeSet
 
-class SimpleMailboxItemView @JvmOverloads constructor(
+class SimpleFolderItemView @JvmOverloads constructor(
     context: Context,
     attrs: AttributeSet? = null,
     defStyleAttr: Int = 0,
-) : SelectableTextItemView(context, attrs, defStyleAttr)
+) : SelectableTextItemView(context, attrs, defStyleAttr), FolderItemView {
+
+    override var indent = 0
+        set(value) {
+            field = value
+            setIndent()
+        }
+
+    init {
+        attrs?.getIndentAttribute(context)
+    }
+}
