+++ app/src/main/res/layout/fragment_menu_drawer.xml
@@ -215,14 +215,14 @@
                 android:orientation="vertical"
                 android:visibility="gone">
 
-                <com.infomaniak.mail.views.SimpleMenuDrawerItemView
+                <com.infomaniak.mail.views.decoratedTextItemView.SimpleMenuDrawerItemView
                     android:id="@+id/importMails"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
                     app:icon="@drawable/ic_drawer_download"
                     app:text="@string/buttonImportEmails" />
 
-                <com.infomaniak.mail.views.SimpleMenuDrawerItemView
+                <com.infomaniak.mail.views.decoratedTextItemView.SimpleMenuDrawerItemView
                     android:id="@+id/restoreMails"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
@@ -236,14 +236,14 @@
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content" />
 
-            <com.infomaniak.mail.views.SimpleMenuDrawerItemView
+            <com.infomaniak.mail.views.decoratedTextItemView.SimpleMenuDrawerItemView
                 android:id="@+id/feedback"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 app:icon="@drawable/ic_feedback"
                 app:text="@string/buttonFeedback" />
 
-            <com.infomaniak.mail.views.SimpleMenuDrawerItemView
+            <com.infomaniak.mail.views.decoratedTextItemView.SimpleMenuDrawerItemView
                 android:id="@+id/help"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
